requirements.txt
Django~=4.1.5
Pillow==9.4.0
asgiref==3.6.0
sqlparse==0.4.3
django-filter==22.1